+++ packages/creamy_field/lib/src/decoration/horizontal_scrollable.dart
@@ -1,4 +1,6 @@
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
 
 class HorizontalScrollable extends StatelessWidget {
   final bool beScrollable;
@@ -6,6 +8,7 @@ class HorizontalScrollable extends StatelessWidget {
   final ScrollController scrollController;
   final ScrollPhysics physics;
   final double horizontalScrollExtent;
+  final bool useExpanded;
 
   /// Scroll padding of the text field under this scrollable
   /// defaults to `const EdgeInsets.only(left: 4)`
@@ -18,28 +21,29 @@ class HorizontalScrollable extends StatelessWidget {
       this.child,
       this.beScrollable = true,
       this.horizontalScrollExtent = 2000,
-      this.padding})
+      this.padding,
+      this.useExpanded})
       : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     if (!beScrollable) return child;
-    return Expanded(
-      child: Padding(
-        padding: padding ?? const EdgeInsets.only(left: 4),
-        // To make child scrollable
-        child: SingleChildScrollView(
-          scrollDirection: Axis.horizontal,
-          controller: scrollController,
-          physics: physics ?? const ClampingScrollPhysics(),
-          child: ConstrainedBox(
-            constraints: BoxConstraints.expand(
-              width: horizontalScrollExtent ?? 2000,
-            ),
-            child: child,
-          ),
+    // widget.padding
+    Widget _scrollable = SingleChildScrollView(
+      scrollDirection: Axis.horizontal,
+      controller: scrollController,
+      physics: physics ?? const ClampingScrollPhysics(),
+      dragStartBehavior: DragStartBehavior.down,
+      child: ConstrainedBox(
+        constraints: BoxConstraints.expand(
+          width: horizontalScrollExtent ?? 2000,
         ),
+        child: child,
       ),
+    );
+    if (!useExpanded) return _scrollable;
+    return Expanded(
+      child: _scrollable,
     );
   }
 }